feat: add compiler and parser with ast
Add a facade for the compiler, you can call it on a file and it will
parse it and show the ast (be sure to use the flags correctly).

Add a very simple(yet not so right) parser for the initial statements.
Implement classes that can be used with the parser to generate ast.

You can run the compiler on the `test.uc` with `python3 uc_compiler.py
test.uc -at-susy`.